% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR.rr-package.R
\docType{package}
\name{MR.rr-package}
\alias{MR.rr}
\alias{MR.rr-package}
\title{MR.rr: Reduced-Rank Mendelian Randomization with Multiple Exposures and Outcomes}
\description{
Implements reduced-rank Mendelian randomization methods for estimating a low-rank causal effect matrix from summary-level genetic association data with multiple exposures and multiple outcomes. Methods include naive, measurement-error-corrected, spectrally regularized, and sparse estimators, together with rank selection and post-selection refitting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MatthewYuexiangPeng/MR.rr}
  \item Report bugs at \url{https://github.com/MatthewYuexiangPeng/MR.rr/issues}
}

}
\author{
\strong{Maintainer}: Yuexiang Peng \email{peng.1276@osu.edu}

Authors:
\itemize{
  \item Zhihong Zhang \email{zhilongz@tamu.edu}
}

}
\keyword{internal}
